server media myself

+++ account/models.py
@@ -5,19 +5,25 @@ from django.utils import timezone
 from .util import getUserCode
 from django.dispatch import receiver
 from project.snowflake import getAccountAUserSnowflakeID
+from hashlib import md5
 import os
-
-
+from UtilGlobal.validator.FieldFile import imagefile_validator
+from media.models import MediaFileSystemStorage
 
 # Create your models here.
     
 
 class AUser(User):
+    @staticmethod
+    def md5_filename(n):
+        s = os.path.splitext(n)
+        return md5(s[0].encode('utf-8')).hexdigest() + s[1]
+
     def cover_dir_path(instance, filename):
-        return f'cover/user-{instance.urlCode}/' + instance.date_joined.strftime('%Y-%m-%d/' + filename)
+        return f'usercover/{instance.urlCode}/' + AUser.md5_filename(filename)
 
     def avatar_dir_path(instance, filename):
-        return f'avatar/user-{instance.urlCode}/' + instance.date_joined.strftime('%Y-%m-%d/' + filename)
+        return f'avatar/{instance.urlCode}/' + AUser.md5_filename(filename)
     
     name = models.CharField(max_length=20,default='some_user')
     # username = models.CharField(  #只能包含_和数字 开头不能有数字或_ _不能2连 结尾不能_
@@ -39,15 +45,15 @@ class AUser(User):
     age = models.PositiveSmallIntegerField(default=None,blank=True,null=True)
     info = models.FileField(default=None,blank=True,null=True)
     additional = models.JSONField(default=None, blank=True, null=True)
-    cover = models.ImageField(upload_to=cover_dir_path, blank=True, default='')
-    avatar = models.ImageField(upload_to=avatar_dir_path, blank=True, default='default/user/user.svg')
+    cover = models.ImageField(upload_to=cover_dir_path, blank=True, default='', validators=[imagefile_validator], storage=MediaFileSystemStorage)
+    avatar = models.ImageField(upload_to=avatar_dir_path, blank=True, default='default/user/user.svg',validators=[imagefile_validator], storage=MediaFileSystemStorage)
 
     def __str__(self) -> str:
         return f"{self.username}({self.urlCode})"
     
     class Meta:
         ordering = ['urlCode']
-    
+
 
     
 
@@ -88,5 +94,5 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
         old_file = old_files[i]
         new_file = new_files[i]
         if not old_file == new_file:
-            if os.path.isfile(old_file.path):
+            if old_file and os.path.isfile(old_file.path):
                 os.remove(old_file.path)
